requirements.txt
binance-futures-connector==4.0.0
rich==13.7.1
python-dotenv==1.0.1
